refactor: Refactor data scratch-ipblocks
