+++ test_runner/src/main/kotlin/ftl/adapter/GoogleIpBlockFetch.kt
@@ -1,0 +1,10 @@
+package ftl.adapter
+
+import ftl.adapter.environment.ipBlocksList
+import ftl.data.IpBlock
+
+object GoogleIpBlockFetch :
+    IpBlock.Fetch,
+    () -> List<IpBlock> by {
+        ipBlocksList()
+    }
