+++ test_runner/src/main/kotlin/ftl/domain/ListIPBlocks.kt
@@ -1,10 +1,12 @@
 package ftl.domain
 
 import flank.common.logLn
-import ftl.environment.ipBlocksListAsTable
+import ftl.adapter.environment.toCliTable
+import ftl.data.fetchIpBLocks
 
 interface ListIPBlocks
 
 operator fun ListIPBlocks.invoke() {
-    logLn(ipBlocksListAsTable())
+    // TODO move toCliTable() to presentation layer during refactor of presentation after #1728
+    logLn(fetchIpBLocks().toCliTable())
 }
