+++ test_runner/src/main/kotlin/ftl/environment/ListLocales.kt
@@ -1,6 +1,11 @@
 package ftl.environment
 
 import com.google.testing.model.Locale
+import ftl.adapter.environment.TAGS
+import ftl.adapter.environment.TestEnvironmentInfo
+import ftl.adapter.environment.createTableColumnFor
+import ftl.adapter.environment.getOrCreateList
+import ftl.adapter.environment.tagToSystemOutColorMapper
 import ftl.util.applyColorsUsing
 import ftl.util.buildTable
