+++ test_runner/src/main/kotlin/ftl/environment/android/ListAndroidDevices.kt
@@ -1,21 +1,21 @@
 package ftl.environment.android
 
 import com.google.testing.model.AndroidModel
-import ftl.environment.EMULATOR_DEVICE
-import ftl.environment.FORM
-import ftl.environment.MAKE
-import ftl.environment.MODEL_ID
-import ftl.environment.MODEL_NAME
-import ftl.environment.OS_VERSION_IDS
-import ftl.environment.PHYSICAL_DEVICE
-import ftl.environment.RESOLUTION
-import ftl.environment.TAGS
-import ftl.environment.TestEnvironmentInfo
-import ftl.environment.VIRTUAL_DEVICE
-import ftl.environment.createTableColumnFor
-import ftl.environment.getOrCreateList
-import ftl.environment.orUnknown
-import ftl.environment.tagToSystemOutColorMapper
+import ftl.adapter.environment.EMULATOR_DEVICE
+import ftl.adapter.environment.FORM
+import ftl.adapter.environment.MAKE
+import ftl.adapter.environment.MODEL_ID
+import ftl.adapter.environment.MODEL_NAME
+import ftl.adapter.environment.OS_VERSION_IDS
+import ftl.adapter.environment.PHYSICAL_DEVICE
+import ftl.adapter.environment.RESOLUTION
+import ftl.adapter.environment.TAGS
+import ftl.adapter.environment.TestEnvironmentInfo
+import ftl.adapter.environment.VIRTUAL_DEVICE
+import ftl.adapter.environment.createTableColumnFor
+import ftl.adapter.environment.getOrCreateList
+import ftl.adapter.environment.orUnknown
+import ftl.adapter.environment.tagToSystemOutColorMapper
 import ftl.util.SystemOutColor
 import ftl.util.applyColorsUsing
 import ftl.util.buildTable
