+++ test_runner/src/main/kotlin/ftl/environment/ios/ListIOsDevices.kt
@@ -1,16 +1,16 @@
 package ftl.environment.ios
 
 import com.google.testing.model.IosModel
-import ftl.environment.MODEL_ID
-import ftl.environment.MODEL_NAME
-import ftl.environment.OS_VERSION_IDS
-import ftl.environment.RESOLUTION
-import ftl.environment.TAGS
-import ftl.environment.TestEnvironmentInfo
-import ftl.environment.createTableColumnFor
-import ftl.environment.getOrCreateList
-import ftl.environment.orUnknown
-import ftl.environment.tagToSystemOutColorMapper
+import ftl.adapter.environment.MODEL_ID
+import ftl.adapter.environment.MODEL_NAME
+import ftl.adapter.environment.OS_VERSION_IDS
+import ftl.adapter.environment.RESOLUTION
+import ftl.adapter.environment.TAGS
+import ftl.adapter.environment.TestEnvironmentInfo
+import ftl.adapter.environment.createTableColumnFor
+import ftl.adapter.environment.getOrCreateList
+import ftl.adapter.environment.orUnknown
+import ftl.adapter.environment.tagToSystemOutColorMapper
 import ftl.util.applyColorsUsing
 import ftl.util.buildTable
