+++ test_runner/src/main/kotlin/ftl/reports/outcome/BillableMinutes.kt
@@ -1,9 +1,9 @@
 package ftl.reports.outcome
 
 import com.google.api.services.toolresults.model.Step
+import ftl.adapter.environment.orUnknown
 import ftl.android.AndroidCatalog
 import ftl.android.DeviceType
-import ftl.environment.orUnknown
 import ftl.util.billableMinutes
 import kotlin.math.min
