+++ test_runner/src/test/kotlin/ftl/environment/TestEnvironmentInfoTest.kt
@@ -1,6 +1,7 @@
 package ftl.environment
 
 import com.google.common.truth.Truth.assertThat
+import ftl.adapter.environment.tagToSystemOutColorMapper
 import ftl.util.SystemOutColor
 import org.junit.Test
